(feat): allow user to checkout cart items

+++ src/main/java/dev/tushar/ecommerceapi/dto/request/CartItemRequestDTO.java
@@ -9,5 +9,5 @@ public record CartItemRequestDTO(
 
         @NotNull(message = "Quantity is required")
         @Min(value = 1, message = "Quantity must be at least 1")
-        Long quantity
+        Integer quantity
 ) {}
